Comprehensive test suite + 80% coverage gate
Mock RPC server (httptest) validates every authenticated request by
recomputing the expected actkey from the request's own params — and parses the
query splitting ONLY on '&' so the ';' literals inside the add= field (srfetch,
dlfails, fetchqueue) survive, matching the real Java server. A mismatch fails
loudly, which is exactly the failure mode that locks accounts.

Coverage (internal/hath): 84.6% total, protocol-critical paths ~100%:
- Auth formulas pinned to independent shasum(1) vectors (rpc_test.go)
- All RPC actions: server_stat, client_login, client_settings, client_start,
  suspend/resume/stop, still_alive, srfetch, dlfetch, dlfails, overload,
  get_blacklist, get_cert, fetchqueue; KEY_EXPIRED clock-resync retry; RPC
  failover
- End-to-end integration: full client + real PKCS#12 + live TLS edge server
  against the mock (Run/cycle/startServer/doShutdown/notifyStart)
- Edge routing: /h cached+proxy+corrupt-reject, /servercmd valid+badkey+
  unauthorized-IP+malformed, /t valid+bad, flood control, admit gating
- Cache: import/delete/lookup, age-weighted prune, startup cleanup+init passes,
  --verify-cache corruption purge, persistent round-trip, blacklist
- Gallery: metadata parse, title sanitize/truncate, full download loop with
  SHA-1 verify + failure reporting
- Proxy: Hath-Request header, SHA-1 verify before serve, cache import
- Bandwidth throttle, settings parse, stats, cert expiry/p12 round-trip

Also fixes a real bug surfaced by tests: DownloadToFile now rejects non-2xx
origin responses (404 was silently treated as a successful tiny download;
Java throws FileNotFoundException).

captureproxy dev tool added to record plain-HTTP RPC traffic from a live Java
instance for mock validation. Makefile: test/vet/cover/gate/ci targets.

+++ internal/hath/settings.go
@@ -35,7 +35,7 @@ type Settings struct {
 	StaticRangeCount int
 
 	// rpc
-	RPCServerPort int16
+	RPCServerPort int
 	RPCPath       string
 	StaticRanges  map[string]bool
 
@@ -210,7 +210,7 @@ func (s *Settings) applySetting(name, value string) {
 		s.SetServerTime(atoi64(value))
 		Debug("setting altered", "server_time_delta", s.serverTimeDelta)
 	case "rpc_server_port":
-		s.RPCServerPort = int16(atoi(value))
+		s.RPCServerPort = atoi(value)
 	case "rpc_server_ip":
 		s.setRPCServers(value)
 	case "rpc_path":
@@ -401,7 +401,7 @@ func (s *Settings) ClearRPCServerFailure() {
 	}
 }
 
-func withPort(host string, port int16) string {
+func withPort(host string, port int) string {
 	if port == 0 || port == 80 {
 		return host
 	}
